Move refreshtoken to abstract parent class
