+++ src/auth/AuthServer.ts
@@ -9,25 +9,14 @@ import {
 import { IAuthService } from './types';
 import HTTP from '../HTTP';
 import { isTokenExpired, jwtToBitloopsUser, parseJwt } from '../helpers';
+import AuthBase from './AuthBase';
 
 type ServerParams = {
   requestParams?: any;
   redirectUrl?: string;
 };
 
-class AuthServer implements IAuthService {
-  private bitloopsConfig: BitloopsConfig;
-
-  private storage: IInternalStorage;
-
-  private http: HTTP;
-
-  constructor(storage: IInternalStorage, http: HTTP, bitloopsConfig: BitloopsConfig) {
-    this.bitloopsConfig = bitloopsConfig;
-    this.storage = storage;
-    this.http = http;
-  }
-
+class AuthServer extends AuthBase implements IAuthService {
   authenticateWithGitHub(): Promise<void> {
     throw new Error('Method not implemented.');
   }
@@ -111,6 +100,7 @@ class AuthServer implements IAuthService {
     // console.log('isRefreshTokenExpired', isRefreshTokenExpired);
     // console.log('isAccessTokenExpired', isAccessTokenExpired);
     if (isRefreshTokenExpired) {
+      console.log('refresh token expired');
       await this.storage.deleteUser();
       return null;
     }
@@ -203,57 +193,9 @@ class AuthServer implements IAuthService {
     const { data } = response.data;
     // console.log('workflow responded', data);
     const { access_token: accessToken, refresh_token: refreshToken } = data;
-    const user = {
-      accessToken,
-      refreshToken,
-    } as BitloopsUser;
+    const jwtData = parseJwt(accessToken);
+    const user = jwtToBitloopsUser(jwtData, accessToken, refreshToken, config.auth.providerId);
     this.storage.saveUser(user);
-  }
-
-  // TODO make parent abstract class and move this func there
-  /**
-   * Tries to refresh token, token must be signed for our clientId,
-   * and not expired for success
-   */
-  async refreshToken(): Promise<BitloopsUser> {
-    const { bitloopsConfig: config } = this;
-    const url = `${config?.ssl === false ? 'http' : 'https'}://${
-      config?.server
-    }/bitloops/auth/refreshToken`;
-    const user = await this.storage.getUser();
-
-    if (config.auth?.authenticationType !== AuthTypes.User) {
-      throw new Error('Attempt to refresh token for non BitloopsUser AuthType');
-    }
-    if (!user?.refreshToken) throw new Error('no refresh token');
-    const body = {
-      refreshToken: user.refreshToken,
-      clientId: config?.auth.clientId,
-      providerId: config?.auth.providerId,
-    };
-    const { data: response, error } = await this.http.handlerWithoutRetries({
-      url,
-      method: 'POST',
-      data: body,
-    });
-    if (error) {
-      console.log('Refresh token was invalid');
-      // invalid refresh token
-      // clean refresh_token
-      // logout user
-      this.clearAuthentication();
-      return Promise.reject(error);
-    }
-    const newAccessToken = response?.data?.accessToken;
-    const newRefreshToken = response?.data?.refreshToken;
-    const newUser: BitloopsUser = {
-      ...user,
-      accessToken: newAccessToken,
-      refreshToken: newRefreshToken,
-    };
-    console.log('Updated refresh token');
-    await this.storage.saveUser(newUser);
-    return newUser;
   }
 }
