+++ src/helpers.ts
@@ -10,6 +10,7 @@ export const wait = (ms: number) =>
 export const isBrowser = () => typeof window !== 'undefined';
 
 export const parseJwt = (token: string): JWTData => {
+  console.log('parseJwt:', token);
   const jwtPayload = token.split('.')[1];
   const base64Payload = jwtPayload.replace(/-/g, '+').replace(/_/g, '/');
   const jsonPayload = decodeURIComponent(
@@ -32,7 +33,7 @@ export const parseJwt = (token: string): JWTData => {
 export const isTokenExpired = (token: string): boolean => {
   const jwtData = parseJwt(token) as JWTData;
   const { exp } = jwtData;
-  // console.log('expires at: ', new Date(exp * 1000));
+  console.log('expires at: ', new Date(exp * 1000));
   const isExpired = Date.now() >= exp * 1000;
   return isExpired;
 };
